+++ src/compiler/symtab.c
@@ -278,6 +278,9 @@ void symtab_init(uint32_t capacity)
 	builtin_list[BUILTIN_SET_ROUNDING_MODE] = KW_DEF("set_rounding_mode");
 	builtin_list[BUILTIN_SIN] = KW_DEF("sin");
 	builtin_list[BUILTIN_STR_HASH] = KW_DEF("str_hash");
+	builtin_list[BUILTIN_STR_UPPER] = KW_DEF("str_upper");
+	builtin_list[BUILTIN_STR_LOWER] = KW_DEF("str_lower");
+	builtin_list[BUILTIN_STR_FIND] = KW_DEF("str_find");
 	builtin_list[BUILTIN_SWIZZLE] = KW_DEF("swizzle");
 	builtin_list[BUILTIN_SWIZZLE2] = KW_DEF("swizzle2");
 	builtin_list[BUILTIN_SQRT] = KW_DEF("sqrt");